database added localized

+++ app/src/main/java/com/example/xenya/openweather/entities/Wind.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Wind(
         @SerializedName("deg")
-        var deg: Int? = 0,
+        var deg: Double? = 0.0,
         @SerializedName("speed")
-        var speed: Int? = 0
+        var speed: Double? = 0.0
 )
